requirements.txt
dagster==1.9.*
dagster-webserver==1.9.*
dagster-postgres==0.25.*

psycopg2-binary==2.9.*
mysql-connector-python==8.3.*
fdb==2.0.2
redis==5.0.*
requests==2.31.*
tenacity==8.2.*
lz4==4.3.*
